refactor(tui): split console run steps
